+++ payments-core/src/main/java/com/stripe/android/StripeIntentResult.kt
@@ -91,6 +91,7 @@ abstract class StripeIntentResult<out T : StripeIntent> internal constructor(
             StripeIntent.NextActionType.AlipayRedirect,
             StripeIntent.NextActionType.WeChatPayRedirect,
             StripeIntent.NextActionType.CashAppRedirect,
+            StripeIntent.NextActionType.SwishRedirect,
             null -> {
                 false
             }