+++ payments-core/src/main/java/com/stripe/android/model/PaymentMethodCreateParams.kt
@@ -34,6 +34,7 @@ data class PaymentMethodCreateParams internal constructor(
     private val usBankAccount: USBankAccount? = null,
     private val link: Link? = null,
     private val cashAppPay: CashAppPay? = null,
+    private val swish: Swish? = null,
     val billingDetails: PaymentMethod.BillingDetails? = null,
     private val metadata: Map<String, String>? = null,
     private val productUsage: Set<String> = emptySet(),
@@ -65,6 +66,7 @@ data class PaymentMethodCreateParams internal constructor(
         usBankAccount: USBankAccount? = null,
         link: Link? = null,
         cashAppPay: CashAppPay? = null,
+        swish: Swish? = null,
         billingDetails: PaymentMethod.BillingDetails? = null,
         metadata: Map<String, String>? = null,
         productUsage: Set<String> = emptySet(),
@@ -84,6 +86,7 @@ data class PaymentMethodCreateParams internal constructor(
         usBankAccount,
         link,
         cashAppPay,
+        swish,
         billingDetails,
         metadata,
         productUsage,
@@ -224,6 +227,17 @@ data class PaymentMethodCreateParams internal constructor(
         metadata = metadata,
     )
 
+    private constructor(
+        swish: Swish,
+        billingDetails: PaymentMethod.BillingDetails?,
+        metadata: Map<String, String>?,
+    ) : this(
+        type = PaymentMethod.Type.Swish,
+        swish = swish,
+        billingDetails = billingDetails,
+        metadata = metadata,
+    )
+
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     fun requiresMandate(): Boolean {
         return requiresMandate
@@ -263,6 +277,11 @@ data class PaymentMethodCreateParams internal constructor(
                 mapOf(code to it)
             }.orEmpty()
         }
+
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    fun cardLast4(): String? {
+        return ((toParamMap()["card"] as? Map<*, *>?)?.get("number") as? String)?.takeLast(4)
+    }
 
     @Parcelize
     data class Card
@@ -490,6 +509,14 @@ data class PaymentMethodCreateParams internal constructor(
      */
     @Parcelize
     class CashAppPay : StripeParamsModel, Parcelable {
+        override fun toParamMap(): Map<String, Any> = emptyMap()
+    }
+
+    /**
+     * Encapsulates parameters used to create [PaymentMethodCreateParams] when using Swish.
+     */
+    @Parcelize
+    class Swish : StripeParamsModel, Parcelable {
         override fun toParamMap(): Map<String, Any> = emptyMap()
     }
 
@@ -956,6 +983,19 @@ data class PaymentMethodCreateParams internal constructor(
             metadata: Map<String, String>? = null
         ): PaymentMethodCreateParams {
             return PaymentMethodCreateParams(CashAppPay(), billingDetails, metadata)
+        }
+
+        /**
+         * Helper method to create [PaymentMethodCreateParams] with [Swish] as the payment
+         * method type.
+         */
+        @JvmStatic
+        @JvmOverloads
+        fun createSwish(
+            billingDetails: PaymentMethod.BillingDetails? = null,
+            metadata: Map<String, String>? = null
+        ): PaymentMethodCreateParams {
+            return PaymentMethodCreateParams(Swish(), billingDetails, metadata)
         }
 
         @JvmStatic