+++ payments-core/src/main/java/com/stripe/android/payments/PaymentFlowResultProcessor.kt
@@ -155,8 +155,13 @@ internal sealed class PaymentFlowResultProcessor<T : StripeIntent, out S : Strip
         val shouldRefreshForCashApp = stripeIntent.requiresAction() &&
             stripeIntent.paymentMethod?.type == PaymentMethod.Type.CashAppPay
 
+        // For Swish, the intent status can still be `requires_action` by the time the user
+        // gets back to the merchant app. We poll until it's succeeded.
+        val shouldRefreshForSwish = stripeIntent.requiresAction() &&
+            stripeIntent.paymentMethod?.type == PaymentMethod.Type.Swish
+
         return succeededMaybeRefresh || cancelledMaybeRefresh ||
-            actionNotProcessedMaybeRefresh || shouldRefreshForCashApp
+            actionNotProcessedMaybeRefresh || shouldRefreshForCashApp || shouldRefreshForSwish
     }
 
     private fun determineFlowOutcome(intent: StripeIntent, originalFlowOutcome: Int): Int {