+++ payments-core/src/test/java/com/stripe/android/view/CardNumberEditTextTest.kt
@@ -3,6 +3,8 @@ package com.stripe.android.view
 import android.text.TextWatcher
 import android.view.ViewGroup
 import androidx.appcompat.view.ContextThemeWrapper
+import androidx.lifecycle.ViewModelStore
+import androidx.lifecycle.ViewModelStoreOwner
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.ApiKeyFixtures
@@ -33,10 +35,12 @@ import com.stripe.android.cards.StaticCardAccountRanges
 import com.stripe.android.core.networking.AnalyticsRequest
 import com.stripe.android.core.networking.AnalyticsRequestExecutor
 import com.stripe.android.model.AccountRange
+import com.stripe.android.model.BinRange
 import com.stripe.android.model.CardBrand
 import com.stripe.android.networking.PaymentAnalyticsRequestFactory
 import com.stripe.android.testharness.ViewTestUtils
 import com.stripe.android.utils.TestUtils.idleLooper
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
@@ -50,7 +54,6 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.LooperMode
 import java.util.concurrent.TimeUnit
 import kotlin.test.AfterTest
 import kotlin.test.Test
@@ -61,7 +64,6 @@ import kotlin.test.assertNull
  * Test class for [CardNumberEditText].
  */
 @RunWith(RobolectricTestRunner::class)
-@LooperMode(LooperMode.Mode.PAUSED)
 internal class CardNumberEditTextTest {
     private val testDispatcher = UnconfinedTestDispatcher()
     private val context = ContextThemeWrapper(
@@ -134,8 +136,8 @@ internal class CardNumberEditTextTest {
     @Test
     fun calculateCursorPosition_whenAmEx_increasesIndexWhenGoingPastTheSpaces() =
         runTest {
-            cardNumberEditText.accountRangeService.updateAccountRangeResult(
-                AccountRangeFixtures.AMERICANEXPRESS
+            cardNumberEditText.accountRangeService.updateAccountRangesResult(
+                listOf(AccountRangeFixtures.AMERICANEXPRESS)
             )
 
             assertThat(
@@ -149,8 +151,8 @@ internal class CardNumberEditTextTest {
     @Test
     fun calculateCursorPosition_whenDinersClub16_decreasesIndexWhenDeletingPastTheSpaces() =
         runTest {
-            cardNumberEditText.accountRangeService.updateAccountRangeResult(
-                AccountRangeFixtures.DINERSCLUB16
+            cardNumberEditText.accountRangeService.updateAccountRangesResult(
+                listOf(AccountRangeFixtures.DINERSCLUB16)
             )
 
             assertThat(
@@ -167,8 +169,8 @@ internal class CardNumberEditTextTest {
     @Test
     fun calculateCursorPosition_whenDeletingNotOnGaps_doesNotDecreaseIndex() =
         runTest {
-            cardNumberEditText.accountRangeService.updateAccountRangeResult(
-                AccountRangeFixtures.DINERSCLUB14
+            cardNumberEditText.accountRangeService.updateAccountRangesResult(
+                listOf(AccountRangeFixtures.DINERSCLUB14)
             )
 
             assertThat(
@@ -179,8 +181,8 @@ internal class CardNumberEditTextTest {
     @Test
     fun calculateCursorPosition_whenAmEx_decreasesIndexWhenDeletingPastTheSpaces() =
         runTest {
-            cardNumberEditText.accountRangeService.updateAccountRangeResult(
-                AccountRangeFixtures.AMERICANEXPRESS
+            cardNumberEditText.accountRangeService.updateAccountRangesResult(
+                listOf(AccountRangeFixtures.AMERICANEXPRESS)
             )
 
             assertThat(
@@ -194,8 +196,8 @@ internal class CardNumberEditTextTest {
     @Test
     fun calculateCursorPosition_whenSelectionInTheMiddle_increasesIndexOverASpace() =
         runTest {
-            cardNumberEditText.accountRangeService.updateAccountRangeResult(
-                AccountRangeFixtures.VISA
+            cardNumberEditText.accountRangeService.updateAccountRangesResult(
+                listOf(AccountRangeFixtures.VISA)
             )
 
             assertThat(
@@ -926,6 +928,114 @@ internal class CardNumberEditTextTest {
         verify(textChangeListener, times(1)).afterTextChanged(any())
     }
 
+    @Test
+    fun `Sets card brand and possible brands correctly when multiple brands without CBC eligibility`() {
+        val accountRanges = listOf(
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.CartesBancaires,
+            ),
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.Visa,
+            ),
+        )
+
+        testAccountRangeService(
+            accountRanges = accountRanges,
+            isCbcEligible = false,
+            dispatcher = testDispatcher,
+        ) { cardNumberEditText ->
+            assertThat(cardNumberEditText.cardBrand).isEqualTo(CardBrand.Unknown)
+            assertThat(cardNumberEditText.possibleCardBrands).isEmpty()
+        }
+    }
+
+    @Test
+    fun `Sets card brand and possible brands correctly when single brand without CBC eligibility`() {
+        val accountRanges = listOf(
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.Visa,
+            ),
+        )
+
+        testAccountRangeService(
+            accountRanges = accountRanges,
+            isCbcEligible = false,
+            dispatcher = testDispatcher,
+        ) { cardNumberEditText ->
+            assertThat(cardNumberEditText.cardBrand).isEqualTo(CardBrand.Visa)
+            assertThat(cardNumberEditText.possibleCardBrands).isEmpty()
+        }
+    }
+
+    @Test
+    fun `Sets card brand and possible brands correctly when multiple brands with CBC eligibility`() = runTest {
+        val accountRanges = listOf(
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.CartesBancaires,
+            ),
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.Visa,
+            ),
+        )
+
+        testAccountRangeService(
+            accountRanges = accountRanges,
+            isCbcEligible = true,
+            dispatcher = testDispatcher,
+        ) { cardNumberEditText ->
+            assertThat(cardNumberEditText.cardBrand).isEqualTo(CardBrand.Unknown)
+            assertThat(cardNumberEditText.possibleCardBrands).containsExactly(CardBrand.CartesBancaires, CardBrand.Visa)
+        }
+    }
+
+    @Test
+    fun `Sets card brand and possible brands correctly when single brand with CBC eligibility`() = runTest {
+        val accountRanges = listOf(
+            AccountRange(
+                binRange = BinRange(
+                    low = "4000000000000000",
+                    high = "4999999999999999",
+                ),
+                panLength = 16,
+                brandInfo = AccountRange.BrandInfo.Visa,
+            ),
+        )
+
+        testAccountRangeService(
+            accountRanges = accountRanges,
+            isCbcEligible = true,
+            dispatcher = testDispatcher,
+        ) { cardNumberEditText ->
+            assertThat(cardNumberEditText.cardBrand).isEqualTo(CardBrand.Visa)
+            assertThat(cardNumberEditText.possibleCardBrands).containsExactly(CardBrand.Visa)
+        }
+    }
+
     private fun verifyCardBrandBin(
         cardBrand: CardBrand,
         bin: String
@@ -940,6 +1050,55 @@ internal class CardNumberEditTextTest {
     private fun updateCardNumberAndIdle(cardNumber: String) {
         cardNumberEditText.setText(cardNumber)
         idleLooper()
+    }
+
+    private fun testAccountRangeService(
+        accountRanges: List<AccountRange>,
+        isCbcEligible: Boolean,
+        dispatcher: CoroutineDispatcher,
+        block: (CardNumberEditText) -> Unit,
+    ) {
+        PaymentConfiguration.init(context, ApiKeyFixtures.FAKE_PUBLISHABLE_KEY)
+
+        val cardWidgetViewModel = CardWidgetViewModel(
+            cbcEnabled = { isCbcEligible },
+            dispatcher = dispatcher,
+        )
+
+        val store = ViewModelStore().apply {
+            val className = CardWidgetViewModel::class.java.canonicalName
+            put("androidx.lifecycle.ViewModelProvider.DefaultKey:$className", cardWidgetViewModel)
+        }
+
+        val storeOwner = object : ViewModelStoreOwner {
+            override val viewModelStore: ViewModelStore = store
+        }
+
+        val activityScenario = activityScenarioFactory.createAddPaymentMethodActivity()
+
+        activityScenario.use { scenario ->
+            scenario.onActivity { activity ->
+                val cardNumberEditText = CardNumberEditText(
+                    context = activity,
+                    workContext = testDispatcher,
+                    cardAccountRangeRepository = DelayedCardAccountRangeRepository(),
+                    analyticsRequestExecutor = analyticsRequestExecutor,
+                    paymentAnalyticsRequestFactory = analyticsRequestFactory,
+                    viewModelStoreOwner = storeOwner,
+                )
+
+                activity.findViewById<ViewGroup>(R.id.add_payment_method_card).also {
+                    it.removeAllViews()
+                    it.addView(cardNumberEditText)
+                }
+
+                testDispatcher.scheduler.advanceTimeBy(2_001)
+
+                cardNumberEditText.accountRangeService.updateAccountRangesResult(accountRanges)
+
+                block(cardNumberEditText)
+            }
+        }
     }
 
     private class DelayedCardAccountRangeRepository : CardAccountRangeRepository {