+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestUpi.kt
@@ -5,7 +5,6 @@ import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.stripe.android.BaseLpmTest
 import com.stripe.android.test.core.Currency
-import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -32,7 +31,6 @@ internal class TestUpi : BaseLpmTest() {
         )
     }
 
-    @Ignore("Investigate the flakiness in nightly tests")
     @Test
     fun testUpiInCustomFlow() {
         testDriver.confirmCustom(