+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## XX.XX.XX - 2023-XX-XX
 
+## 20.32.0 - 2023-09-25
+
+### PaymentSheet
+* [ADDED][7337](https://github.com/stripe/stripe-android/pull/7337) PaymentSheet now supports Swish for PaymentIntents in private beta.
+
+### Payments
+* [ADDED][7337](https://github.com/stripe/stripe-android/pull/7337) Added support for Swish for PaymentIntents in private beta.
+
 ### Financial Connections
 * [FIXED][7331](https://github.com/stripe/stripe-android/pull/7331) When cancelling out of a auth session, going back to the consent screen required two back taps.